pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "architecture-law-agent"
version = "0.1.0"
description = "RAG-based architecture law agent for Korean building law and enforcement decree"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Architecture Team" }
]
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "pandas>=2.2.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "langchain>=1.2.0",
  "langgraph>=1.0.0",
  "langchain-naver>=0.1.0",
  "langchain-qdrant>=0.2.0",
  "qdrant-client>=1.12.0",
  "pdfplumber>=0.11.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/architecture_agent"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/tests"]
addopts = "-q"
